adds a populate({})

+++ model/contacts.js
@@ -1,14 +1,21 @@
 const Contact = require('./schemas/contact');
 
 // GET
-const getAll = async () => {
-  const results = await Contact.find({});
+const getAll = async userId => {
+  const results = await Contact.find({ owner: userId }).populate({
+    // Substants the value and fields of the document instead of ID
+    path: 'owner',
+    select: 'name email gender -_id', // -_id (removes the field from output)
+  });
   return results;
 };
 
 // GET by ID
-const getById = async id => {
-  const result = await Contact.findOne({ _id: id });
+const getById = async (userId, id) => {
+  const result = await Contact.findOne({ _id: id, owner: userId }).populate({
+    path: 'owner',
+    select: 'name email gender -_id',
+  });
   return result;
 };
 
@@ -19,17 +26,15 @@ const create = async body => {
 };
 
 // DELETE
-const remove = async id => {
-  const result = await Contact.findByIdAndRemove({ _id: id });
+const remove = async (userId, id) => {
+  const result = await Contact.findByIdAndRemove({ _id: id, owner: userId });
   return result;
 };
 
 // UPDATE
-const update = async (id, body) => {
+const update = async (userId, id, body) => {
   const result = await Contact.findOneAndUpdate(
-    {
-      _id: id,
-    },
+    { _id: id, owner: userId },
     { ...body },
     { new: true },
   );
